requirements.txt
streamlit
requests
PyGitHub
sentence-transformers
chromadb
openai
langchain
langchain-openai
langchain-community
tiktoken
python-dotenv